pyproject.toml
[project]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
name = "avilla-workspace"
dependencies = [
    "graia-broadcast>=0.23.1",
    "graia-amnesia>=0.8.1",
    "creart>=0.3.0",
    "graia-saya>=0.0.18",
    "launart>=0.8.1",
    "loguru>=0.6.0",
    "aiohttp>=3.8.1",
    "nonebot2>=2.0.1",
    "wsproto>=1.2.0",
    "starlette>=0.30.0",
    "uvicorn>=0.23.1",
    "nonechat<1.0.0,>=0.2.0",
    "arclet-alconna>=1.7.16",
    "arclet-alconna-tools>=0.6.5",
    "pygtrie>=2.5.0",
    "selectolax>=0.3.16",
    "satori-python>=0.5.0",
    "python-telegram-bot>=20.4",
]
version = "0"
requires-python = ">=3.9"

[build-system]
requires = ["mina-build>=0.5.1"]
build-backend = "mina.backend"

[tool.pdm.build]
includes = ["avilla/"]

[tool.mina]
enabled = true

[tool.black]
include = '\.pyi?$'
line-length = 120
target-version = ['py39']

[tool.isort]
profile = "black"

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
    "mina-build>=0.4.0",
    "richuru>=0.1.1",
    "ruff>=0.0.272",
    "uvicorn>=0.22.0",
    "devtools>=0.11",
    "ipykernel>=6.23.2",
    "pip>=23.2.1",
    "pdm-mina>=0.3.1",
]
